+++ Source/Readers/HTKMLFReader/latticearchive.cpp
@@ -154,7 +154,7 @@ static size_t tryfind(const MAPTYPE &map, const KEYTYPE &key, VALTYPE deflt)
     // TODO: This is sort of redundant now--it gets the symmap from the HMM, i.e. always the same for all archives.
     writeunitmap(symlistpath, modelsymmap);
 
-    fprintf(stderr, "completed %lu out of %lu lattices (%lu read failures, %.1f%%)\n", infiles.size(), infiles.size() - brokeninputfiles, brokeninputfiles, 100.0f * brokeninputfiles / infiles.size());
+    fprintf(stderr, "completed %lu out of %lu lattices (%lu read failures, %.1f%%)\n", (unsigned long)infiles.size(), (unsigned long)(infiles.size() - brokeninputfiles), (unsigned long)brokeninputfiles, 100.0f * brokeninputfiles / infiles.size());
 }
 
 // helper to set a context value (left, right) with checking of uniqueness
@@ -204,7 +204,7 @@ vector<lattice::nodecontext> lattice::determinenodecontexts(const msra::asr::sim
                     // ends with n = position of furthest non-sp
                     if (n < 0) // only sp?
                         RuntimeError("determinenodecontexts: word consists only of /sp/");
-                    fprintf(stderr, "determinenodecontexts: word with %lu /sp/ at the end found, edge %d\n", a.size() - 1 - n, j);
+                    fprintf(stderr, "determinenodecontexts: word with %lu /sp/ at the end found, edge %d\n", (unsigned long)a.size() - 1 - n, j);
                     multispseen++;
                     Z = a[n].unit;
                 }
@@ -229,7 +229,7 @@ vector<lattice::nodecontext> lattice::determinenodecontexts(const msra::asr::sim
         nodecontexts[E].setleft(Zid);
     }
     if (multispseen > 0)
-        fprintf(stderr, "determinenodecontexts: %lu broken edges in %lu with multiple /sp/ at the end seen\n", multispseen, edges.size());
+        fprintf(stderr, "determinenodecontexts: %lu broken edges in %lu with multiple /sp/ at the end seen\n", (unsigned long)multispseen, (unsigned long)edges.size());
     // check CI conditions and put in 't'
     // We make the hard assumption that there is only one CI phone, /sil/.
     const auto &silhmm = hset.gethmm(silunit);
@@ -323,7 +323,7 @@ void lattice::merge(const lattice &other, const msra::asr::simplesenonehmm &hset
         if (contexts[i].iother != SIZE_MAX)
             othernodemap[contexts[i].iother] = j - 1;
     }
-    fprintf(stderr, "merge: joint node space uniq'ed to %d from %d\n", j, contexts.size());
+    fprintf(stderr, "merge: joint node space uniq'ed to %d from %d\n", j, (unsigned long)contexts.size());
     contexts.resize(j);
 
     // create a new node array (just copy the contexts[].t fields)
@@ -382,7 +382,7 @@ void lattice::dedup()
         }
         edges2[k++] = edges2[j];
     }
-    fprintf(stderr, "dedup: edges reduced to %d from %d\n", k, edges2.size());
+    fprintf(stderr, "dedup: edges reduced to %lu from %lu\n", (unsigned long)k, (unsigned long)edges2.size());
     edges2.resize(k);
     info.numedges = edges2.size();
     edges.clear(); // (should already be, but isn't; make sure we no longer use it)
@@ -515,13 +515,13 @@ void lattice::dedup()
         }
     } // end for (toclines)
     if (skippedmerges > 0)
-        fprintf(stderr, "convert: %d out of %d merge operations skipped due to secondary lattice missing\n", skippedmerges, toclines.size());
+        fprintf(stderr, "convert: %lu out of %lu merge operations skipped due to secondary lattice missing\n", (unsigned long)skippedmerges, (unsigned long)toclines.size());
 
     // write out the updated unit map
     if (f && ftoc)
         writeunitmap(symlistpath, modelsymmap);
 
-    fprintf(stderr, "converted %d lattices\n", toclines.size());
+    fprintf(stderr, "converted %lu lattices\n", (unsigned long)toclines.size());
 }
 
 // ---------------------------------------------------------------------------
@@ -542,7 +542,7 @@ void lattice::fromhtklattice(const wstring &path, const std::unordered_map<std::
     for (; iter != lines.end() && strncmp(*iter, "N=", 2); iter++)
     {
         if (strncmp(*iter, "lmscale=", 8) == 0) // note: HTK sometimes generates extra garbage space at the end of this line
-            if (sscanf_s(*iter, "lmscale=%f wdpenalty=%f%c", &info.lmf, &info.wp, &dummychar, sizeof(dummychar)) != 2 && dummychar != ' ')
+            if (sscanf_s(*iter, "lmscale=%f wdpenalty=%f%c", &info.lmf, &info.wp, &dummychar, (unsigned int)sizeof(dummychar)) != 2 && dummychar != ' ')
                 RuntimeError("lattice: mal-formed lmscale/wdpenalty line in lattice: %s", *iter);
     }
 
@@ -550,7 +550,7 @@ void lattice::fromhtklattice(const wstring &path, const std::unordered_map<std::
     if (iter != lines.end())
     {
         unsigned long N, L;
-        if (sscanf_s(*iter, "N=%lu L=%lu %c", &N, &L, &dummychar, sizeof(dummychar)) != 2)
+        if (sscanf_s(*iter, "N=%lu L=%lu %c", &N, &L, &dummychar, (unsigned int)sizeof(dummychar)) != 2)
             RuntimeError("lattice: mal-formed N=/L= line in lattice: %s", *iter);
         info.numnodes = N;
         info.numedges = L;
@@ -568,7 +568,7 @@ void lattice::fromhtklattice(const wstring &path, const std::unordered_map<std::
             RuntimeError("lattice: not enough I lines in lattice");
         unsigned long itest;
         float t;
-        if (sscanf_s(*iter, "I=%lu t=%f%c", &itest, &t, &dummychar, sizeof(dummychar)) < 2)
+        if (sscanf_s(*iter, "I=%lu t=%f%c", &itest, &t, &dummychar, (unsigned int)sizeof(dummychar)) < 2)
             RuntimeError("lattice: mal-formed node line in lattice: %s", *iter);
         if (i != (size_t) itest)
             RuntimeError("lattice: out-of-sequence node line in lattice: %s", *iter);
@@ -590,7 +590,7 @@ void lattice::fromhtklattice(const wstring &path, const std::unordered_map<std::
         char d[1024];
         // example:
         // J=12    S=1    E=13   a=-326.81   l=-5.090  d=:sil-t:s+k:e,0.03:dh:m-ax:m+sil,0.03:sil,0.02:
-        int nvals = sscanf_s(*iter, "J=%lu S=%lu E=%lu a=%f l=%f d=%s", &jtest, &S, &E, &a, &l, &d, sizeof(d));
+        int nvals = sscanf_s(*iter, "J=%lu S=%lu E=%lu a=%f l=%f d=%s", &jtest, &S, &E, &a, &l, &d, (unsigned int)sizeof(d));
         if (nvals == 5 && j == info.numedges - 1) // special case: last edge is a !NULL and thus may have the d= record missing
             strcpy(d, ":");
         else if (nvals != 6)
@@ -633,7 +633,7 @@ void lattice::fromhtklattice(const wstring &path, const std::unordered_map<std::
         if (edgeframes != nodes[E].t - (size_t) nodes[S].t)
         {
             char msg[128];
-            sprintf(msg, "\n-- where edgeframes=%d != (nodes[E].t - nodes[S].t=%d), the gap is %d.", edgeframes, nodes[E].t - (size_t) nodes[S].t, edgeframes + nodes[S].t - nodes[E].t);
+            sprintf(msg, "\n-- where edgeframes=%lu != (nodes[E].t - nodes[S].t=%lu), the gap is %lu.", (unsigned long)edgeframes, (unsigned long)(nodes[E].t - (size_t) nodes[S].t), (unsigned long)(edgeframes + nodes[S].t - nodes[E].t));
             RuntimeError("lattice: alignment info duration mismatches edge duration: %s%s", *iter, msg);
         }
     }
